refactor foregroundservice: expose event and make it mockable
